requirements.txt
aiohttp
zmq
ruamel.yaml